finish resize and add resize optimization
