+++ src/core/dom.js
@@ -56,7 +56,9 @@ class Dom {
   }
 
   css(styles = {}) {
-
+    Object.keys(styles).forEach(key => {
+      this.$el.style[key] = styles[key];
+    });
   }
 }
