+++ src/scss/table.scss
@@ -72,6 +72,7 @@
     opacity: 0;
     background: $primary-color;
     border-radius: 20%;
+    z-index: 10;
   }
 
   .col-resize {
@@ -79,7 +80,7 @@
     width: 4px;
     
     &:hover {
-      opacity: 1;
+      opacity: 1!important;
       cursor: col-resize;
     }
   }
@@ -89,7 +90,7 @@
     height: 4px;
 
     &:hover {
-      opacity: 1;
+      opacity: 1!important;
       cursor: row-resize;
     }
   }
